+++ tools/ncurses/samples/initscr.c
@@ -1,0 +1,14 @@
+#include <ncurses.h>
+#include <string.h>
+#include <unistd.h>
+
+int main(void)
+{
+	initscr();
+	cbreak();
+	noecho();
+	curs_set(0);
+
+	endwin();
+	return 0;
+}